+++ src/modules/print/tests/unit/js/getVisibleLayer.spec.js
@@ -40,7 +40,10 @@ describe("src/modules/print/utils/getVisibleLayer.js", function () {
             getMaxResolution: () => 1000,
             getMinResolution: () => 0,
             getZIndex: () => 1,
-            get: () => "maybeInvisible"
+            get: () => "maybeInvisible",
+            addEventListener: sinon.stub(),
+            setZIndex: () => sinon.stub(),
+            setOpacity: () => sinon.stub()
         };
         layer2 = {
             id: "2",
@@ -146,6 +149,21 @@ describe("src/modules/print/utils/getVisibleLayer.js", function () {
             expect(store.commit.firstCall.args[1]).to.deep.equals([]);
             expect(store.commit.secondCall.args[0]).to.equals("Modules/Print/setInvisibleLayerNames");
             expect(store.commit.secondCall.args[1]).to.equals("");
+        });
+
+        it("sets zIndex for child layers of a group", function () {
+            const setZIndexSpy = sinon.spy(),
+                groupLayer = new LayerGroup({
+                    visible: true,
+                    zIndex: 5,
+                    layers: [layer1, layer2]
+                });
+
+            layer2.setZIndex = setZIndexSpy;
+            layers.push(groupLayer);
+            layerProvider.getVisibleLayer();
+
+            expect(setZIndexSpy.calledWith(5)).to.be.true;
         });
 
     });